+++ docs/fr_FR/changelog.md
@@ -4,14 +4,16 @@
 - Créer une documentation
 - Améliorer la sélection de la station favoris (au lieu d'afficher le fichier JSON brut)
 - Pouvoir sur un seul véhicule avoir le choix entre suivre une station favorite et des stations autour des coordonnées
-- se servir du plugin geotrav pour obtenir les coordonées (prendre exemple sur le plugin DarkSky ?)
 - mettre en avant une confiance par rapport à la date de dernière mise à jour.
 
+
+## 12 juin 2020
+- Changement de la sélection de la localisation. Il faut désormais sélectionner la localisation (geotrav ou configuration) plutôt que rentrer les coordonnées.
 
 ## 11 juin 2020
 - Mise à jour de la fonctionnalité de sélection du nombre de station à surveiller, qui ne fonctionnait pas quand l'équipement était créé avant la mise à jour précédente
 - Ajout d'un message quand le nombre de station surveillé à supérieur au nombre de station disponible dans le rayon sélectionné
 
-=======
+
 ## 10 juin 2020
 - Ajout de la possibilité de choisir le nombre de station surveillée, jusqu'à 10 maximum